requirements.txt
requests
anthropic
python-telegram-bot
apscheduler
pandas
python-dotenv
yfinance